party sheet: - use DiceRoller for all rolls; - use RollDialog to display all roll dialogs - add comments to new methods - cleanup
